Default to use no authentication at command-line
When resolving an identifier (semi-)interactively at the command line (as opposed to in server mode), we don’t need API authentication, because no server will be run. Therefore, turn off authentication so we don’t get warnings or an insistence on setting an authentication provider that won’t be used!