Add CORS configuration.
In a real deployment, we will use envoy or an alternative gateway, but we’re running on a tiny linux VM and so for now we make this gRPC server implement some of the functionality of other microservices. This is simply to get a local single-server deployment for use within a private network.